reuse transform/bin reducers cleaner API (Plot.hex is the shortcut, and Plot.hexbin the options transform)
