+++ src/index.js
@@ -15,10 +15,10 @@ export {TickX, TickY, tickX, tickY} from "./marks/tick.js";
 export {Vector, vector} from "./marks/vector.js";
 export {valueof} from "./options.js";
 export {dodgeX, dodgeY} from "./layouts/dodge.js";
+export {hex, hexbin} from "./layouts/hexbin.js";
 export {filter, reverse, sort, shuffle} from "./transforms/basic.js";
 export {bin, binX, binY} from "./transforms/bin.js";
 export {group, groupX, groupY, groupZ} from "./transforms/group.js";
-export {hexbin, hexbinFill, hexbinOpacity, hexbinR, hexbinText} from "./transforms/hexbin.js";
 export {normalize, normalizeX, normalizeY} from "./transforms/normalize.js";
 export {map, mapX, mapY} from "./transforms/map.js";
 export {window, windowX, windowY} from "./transforms/window.js";
